servers/django-strawberry/pyproject.toml
[tool.poetry]
name = "django-graphene"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
strawberry-graphql = "^0.77.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
